This is the 3.4.87 stable release

Conflicts:
	arch/arm/boot/compressed/Makefile
	arch/arm/boot/compressed/head.S
	arch/arm/include/asm/processor.h
	drivers/hid/hid-input.c
	drivers/usb/host/xhci-hub.c
	drivers/usb/serial/qcserial.c
	include/linux/compiler-gcc.h
	include/linux/idr.h
	include/linux/msg.h
	include/linux/pps_kernel.h
	include/linux/usb/serial.h
	lib/idr.c

+++ include/linux/msg.h
@@ -60,9 +60,9 @@ struct msginfo {
 #include <linux/list.h>
 
 struct msg_msg {
-	struct list_head m_list; 
-	long  m_type;          
-	int m_ts;           
+	struct list_head m_list;
+	long m_type;
+	size_t m_ts;		/* message text size */
 	struct msg_msgseg* next;
 	void *security;
